+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Changed
 ### Fixed
 
+* Filter SDL1 active events (#351)
+
 ## [1.4.0] - 2020-06-19
 ### Added